move health check api to the api folder & metric service inside the services folder
