+++ skill-tree/src/main/java/com/RDS/skilltree/services/MetricService.java
@@ -1,0 +1,5 @@
+package com.RDS.skilltree.services;
+
+public interface MetricService {
+    Double getUptime();
+}
